+++ src/outputs/outputs.cpp
@@ -256,10 +256,10 @@ Outputs::Outputs(Mesh *pm, ParameterInput *pin, SimTime *tm) {
                   pin->GetVector<std::string>(pib->block_name, swname + "_variables");
               op.swarms[swname].insert(varnames.begin(), varnames.end());
             }
-            // Always output x, y, and z for swarms so that they work with vis tools.
-            std::vector<std::string> coords = {swarm_position::x::name(),
-                                               swarm_position::y::name(),
-                                               swarm_position::z::name()};
+            // Always output id, x, y, and z for swarms so that they work with vis tools.
+            std::vector<std::string> coords = {
+                swarm_position::id::name(), swarm_position::x::name(),
+                swarm_position::y::name(), swarm_position::z::name()};
             op.swarms[swname].insert(coords.begin(), coords.end());
           }
         }
@@ -292,6 +292,8 @@ Outputs::Outputs(Mesh *pm, ParameterInput *pin, SimTime *tm) {
         }
 #ifdef ENABLE_HDF5
         op.write_xdmf = pin->GetOrAddBoolean(op.block_name, "write_xdmf", true);
+        // TODO(reviewers) Why are do we forcibly disallow writing particle xdmf for
+        // restarts?
         op.write_swarm_xdmf =
             (restart) ? false
                       : pin->GetOrAddBoolean(op.block_name, "write_swarm_xdmf", false);